Adding function to set/get Path* options in PyRosetta. Refactoring PyJobDistributor so it works more reliably on filesystem with caching. Adding option to write compressed PDB files. Adding JD unit tests.

Old repository SHA1: cfc97a69b5f9576c6f606e3d32ad2c4c2a335f8a

+++ source/src/python/PyRosetta/src/pyrosetta/__init__.py
@@ -10,7 +10,7 @@ from __future__ import absolute_import
 ###############################################################################
 # Imports.
 # Standard library.
-import os, sys, platform, os.path, json, random
+import os, sys, platform, os.path, json, random, gzip
 
 import pyrosetta.rosetta as rosetta
 import pyrosetta.method_bindings
@@ -382,9 +382,12 @@ def output_scorefile(pose, pdb_name, current_name, scorefilepath, \
 
 
 class PyJobDistributor:
-    def __init__(self, pdb_name, nstruct, scorefxn):
+    def __init__(self, pdb_name, nstruct, scorefxn, compress=False):
         self.pdb_name = pdb_name
         self.nstruct = nstruct
+        self.compress = compress
+
+        self.current_id = None
         self.current_name = None	      # Current decoy name
         self.scorefxn = scorefxn	      # Used for final score calculation
         self.native_pose = None		      # Used for rmsd calculation
@@ -399,7 +402,9 @@ class PyJobDistributor:
 
     def start_decoy(self):
         while(self.sequence):
-            self.current_name = self.pdb_name + '_' + str(self.sequence.pop()) + '.pdb'
+            self.current_id = self.sequence[0]
+
+            self.current_name = self.pdb_name + '_' + str(self.current_id) + ('.pdb.gz' if self.compress else '.pdb')
             self.current_in_progress_name = self.current_name + '.in_progress'
 
             if (not os.path.isfile(self.current_name))  and  (not os.path.isfile(self.current_in_progress_name)):
@@ -409,16 +414,39 @@ class PyJobDistributor:
 
 
     def output_decoy(self, pose):
-        if os.path.isfile(self.current_in_progress_name):
-            dump_pdb(pose, self.current_name)  # Outputs pdb file.
-            os.remove(self.current_in_progress_name)
+        if os.path.isfile(self.current_name): # decoy is already exist, probably written to other process -> moving to next decoy if any
 
-            score_tag = ".fasc"
-            if not pose.is_fullatom(): score_tag = ".sc"
+            if os.path.isfile(self.current_in_progress_name): os.remove(self.current_in_progress_name)
+
+            if not self.job_complete:
+                self.start_decoy()
+                self.output_decoy(self, pose)
+
+        else:
+            if self.compress:
+                s = pyrosetta.rosetta.std.ostringstream()
+                pose.dump_pdb(s)
+
+                z = pyrosetta.rosetta.utility.io.ozstream(self.current_name)
+                z.write(s.str(), len(s.str() ) )
+                del z
+
+                #s = pyrosetta.rosetta.std.ostringstream()
+                #pose.dump_pdb(s)
+                #with gzip.open(self.current_name, 'wb') as f: f.write( s.str() )
+
+            else:
+                dump_pdb(pose, self.current_name)
+
+            score_tag = '.fasc' if pose.is_fullatom() else '.sc'
 
             scorefile = self.pdb_name + score_tag
             output_scorefile(pose, self.pdb_name, self.current_name, scorefile, self.scorefxn,
                              self.nstruct, self.native_pose, self.additional_decoy_info)
+
+            self.sequence.remove(self.current_id)
+
+            if os.path.isfile(self.current_in_progress_name): os.remove(self.current_in_progress_name)
 
             self.start_decoy()
